+++ lib/memfs/file.rb
@@ -278,6 +278,10 @@ module MemFs
       File.ctime(path)
     end
 
+    def fileno
+      entry.fileno
+    end
+
     def flock(*)
       SUCCESS
     end