+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## HEAD
 
 * ADD: `Dir.empty?` from Ruby 2.4
+* ADD: `File#fileno` and `Dir#fileno` raise `NotImplementedError`
 
 ## 1.0.0